pyproject.toml
[tool.poetry]
name = "lattice-backup"
version = "0.1.0"
description = ""
authors = ["Ester Beltrami <ester.beltrami@torchbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.25.1"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.106.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"


[tool.poetry.group.dev.dependencies]
detect-secrets = "^1.4.0"

[tool.poetry.scripts]
backup = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
